RHELLikeScript unittests
