+++ pkg/scripts/os_test.go
@@ -155,151 +155,3 @@ func TestMigrateToContainerd(t *testing.T) {
 		})
 	}
 }
-
-func TestKubeadmFlatcar(t *testing.T) {
-	t.Parallel()
-
-	type args struct {
-		cluster kubeoneapi.KubeOneCluster
-	}
-	tests := []struct {
-		name string
-		args args
-		err  error
-	}{
-		{
-			name: "force",
-			args: args{
-				cluster: genCluster(),
-			},
-		},
-		{
-			name: "overwrite registry",
-			args: args{
-				cluster: genCluster(
-					withRegistry("127.0.0.1:5000"),
-				),
-			},
-		},
-		{
-			name: "with containerd",
-			args: args{
-				cluster: genCluster(),
-			},
-		},
-		{
-			name: "with containerd with insecure registry",
-			args: args{
-				cluster: genCluster(
-					withInsecureRegistry("127.0.0.1:5000"),
-				),
-			},
-		},
-		{
-			name: "with cilium",
-			args: args{
-				cluster: genCluster(withCiliumCNI),
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			got, err := KubeadmFlatcar(&tt.args.cluster)
-			if !errors.Is(err, tt.err) {
-				t.Errorf("KubeadmFlatcar() error = %v, wantErr %v", err, tt.err)
-
-				return
-			}
-
-			testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-		})
-	}
-}
-
-func TestRemoveBinariesDebian(t *testing.T) {
-	t.Parallel()
-
-	got, err := RemoveBinariesDebian()
-	if err != nil {
-		t.Errorf("RemoveBinariesDebian() error = %v", err)
-
-		return
-	}
-
-	testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-}
-
-func TestRemoveBinariesCentOS(t *testing.T) {
-	t.Parallel()
-
-	got, err := RemoveBinariesCentOS()
-	if err != nil {
-		t.Errorf("RemoveBinariesCentOS() error = %v", err)
-
-		return
-	}
-
-	testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-}
-
-func TestRemoveBinariesAmazonLinux(t *testing.T) {
-	t.Parallel()
-
-	got, err := RemoveBinariesAmazonLinux()
-	if err != nil {
-		t.Errorf("RemoveBinariesAmazonLinux() error = %v", err)
-
-		return
-	}
-
-	testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-}
-
-func TestRemoveBinariesFlatcar(t *testing.T) {
-	t.Parallel()
-
-	got, err := RemoveBinariesFlatcar()
-	if err != nil {
-		t.Errorf("RemoveBinariesFlatcar() error = %v", err)
-
-		return
-	}
-
-	testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-}
-
-func TestUpgradeKubeadmAndCNIFlatcar(t *testing.T) {
-	t.Parallel()
-
-	c := genCluster(
-		withKubeVersion("1.26.0"),
-		withInsecureRegistry("127.0.0.1:5000"),
-	)
-	got, err := UpgradeKubeadmAndCNIFlatcar(&c)
-	if err != nil {
-		t.Errorf("UpgradeKubeadmAndCNIFlatcar() error = %v", err)
-
-		return
-	}
-
-	testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-}
-
-func TestUpgradeKubeletAndKubectlFlatcar(t *testing.T) {
-	t.Parallel()
-
-	c := genCluster(
-		withKubeVersion("1.26.0"),
-		withInsecureRegistry("127.0.0.1:5000"),
-	)
-	got, err := UpgradeKubernetesBinariesFlatcar(&c)
-	if err != nil {
-		t.Errorf("UpgradeKubeletAndKubectlFlatcar() error = %v", err)
-
-		return
-	}
-
-	testhelper.DiffOutput(t, testhelper.FSGoldenName(t), got, *updateFlag)
-}
